Feature/liquity fuse test fixes (#368)
This PR is due to the fact that the current BalanceFuse of Liquity does not take into account unrealized gains both for collateral and for BOLD. They can both change in a typical use case: the BOLD ones simply accrue due to the fees accumulated from the Trove payments, the collateral change in case of a liquidation.

The corrected calculation is therefore

stashedColl + depositorCollGain + compoundedBoldDeposit + getDepositorYieldGain
Besides this functional change, the PR has:

A small refactoring of LiquityBalanceFuse.sol to avoid stack too deep errors and improve readability
Change in naming conventions of variables into variable_ to adjust to IPOR standards
A README.md file to document the Fuse
Improved tests, where totalAssets() and totalAssetsInMarkets(LIQUITY_V2) are checked before and after every execution.
A small fix of the README.md in Ebisu, where there was a mistake in the market id.
---------

Co-authored-by: idrees535 <idrees535@gmail.com>