Ebisu: bug fix for ebisu flashloan amount (#394)
In the Ebisu Zapper, we have
uint256 totalCollAmount = _params.collAmount + _effectiveFlashLoanAmount;
and the Liquity native borrowerOperations.openTrove function is called with totalCollAmount rather than collAmount.
Instead, in our check _checkUpfrontFeeAndDebt() we only consider collAmount, thus the call reverts unnecessarily in case the flashloan amount is higher than the collateral amount (leverage higher than 2).

In the new test we have a higher leverage, and with the modified fuse, the call to Ebisu succeeds. Instead, by reverting the Fuse, the call fails at the check level, thus showcasing the fact that the check is wrong.